fix(api): give duplicated entities their own storage objects
